requirements.txt
flask==2.0.1
requests==2.31.0
schedule==1.1.0
pandas==1.5.3
pyyaml==6.0
apscheduler==3.10.1
python-dotenv==1.0.0
